[fix] Rename newKit directory to kit_v2
