+++ aim/web/ui/src/stories/Button.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import Button from 'components/newKit/Button';
+import Button from 'components/kit_v2/Button';
 
 export default {
   title: 'Kit/Inputs/Button',
